% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_numbers}
\alias{add_numbers}
\title{Export Header-only Function into R}
\usage{
add_numbers(x, y)
}
\arguments{
\item{x}{A \code{numeric vector}.}

\item{y}{A single \code{numeric} value.}
}
\value{
A \code{numeric vector} that is incremented by a single value.
}
\description{
Retrieve a function defined inline from the package's header
and surface it into \emph{R}.
}
\examples{
x = 1:5
y = add_numbers(x, 2)
y
}
