% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppHeaderSharing-package.R
\docType{package}
\name{RcppHeaderSharing-package}
\alias{RcppHeaderSharing}
\alias{RcppHeaderSharing-package}
\title{RcppHeaderSharing: Prototyped Sharing of C++ Inline Headers}
\description{
Provides a set of 'C++' header files with definitions inlined. Functions are then surfaced for 'R' using 'Rcpp Attributes' while the definitions can be used in other 'R' packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rd-rcpp.thecoatlessprofessor.com/rcpp-shared-cpp-functions/}
  \item \url{https://github.com/coatless-rd-rcpp/rcpp-shared-cpp-functions}
  \item Report bugs at \url{https://github.com/coatless-rd-rcpp/rcpp-shared-cpp-functions/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
\keyword{internal}
